# apps/api/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PULL Super App API
  version: 1.0.0
  description: |
    # PULL Super App API

    Complete API for the PULL Super App platform, providing endpoints for:

    - **Authentication**: User registration, login, token refresh, and password management
    - **Trading**: Order management, portfolio tracking, and buying power
    - **Predictions**: Prediction market events, positions, and market data
    - **RWA (Real-World Assets)**: Asset browsing, ownership, and purchases
    - **Rewards**: Points system, leaderboards, and reward redemption

    ## Authentication

    Most endpoints require authentication via Bearer token. Include the token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    Tokens are obtained through the `/api/auth/login` or `/api/auth/register` endpoints.

    ## Rate Limiting

    API requests are rate-limited. Current limits are returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window

    ## Error Handling

    All errors follow a consistent format with a `success: false` flag and an `error` object containing:
    - `code`: Machine-readable error code
    - `message`: Human-readable error description

    ## Request Tracking

    All responses include a `requestId` for debugging and support purposes.
  contact:
    name: PULL Support
    email: support@pull.app
...
  license:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token obtained from login or register endpoints
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
          enum: [false]
          description: Always false for error responses
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error message
              example: Invalid email address
        requestId:
          type: string
          format: uuid
          description: Request tracking ID for debugging
          example: 550e8400-e29b-41d4-a716-446655440000
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp
          example: "2024-01-15T10:30:00.000Z"
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          minLength: 8
          description: Password (min 8 chars, must include uppercase, lowercase, and number)
          example: SecurePass123
        displayName:
          type: string
          minLength: 2
          maxLength: 50
          description: Optional display name
          example: John Doe
        referralCode:
          type: string
          description: Optional referral code from another user
          example: PULL2024ABC
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          description: User password
          example: SecurePass123
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Email address associated with the account
          example: user@example.com
    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
          description: Unique user ID
          example: usr_abc123def456
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        displayName:
          type: string
          description: User display name
          example: John Doe
        username:
          type: string
          description: Unique username
          example: johndoe
        avatarUrl:
          type: string
          format: uri
          description: URL to user avatar image
          example: https://cdn.pull.app/avatars/abc123.jpg
        emailVerified:
          type: boolean
          description: Whether email has been verified
          example: true
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
          description: KYC verification status
          example: approved
        kycTier:
          type: integer
          minimum: 0
          maximum: 3
          description: KYC tier level (0-3)
          example: 2
        referralCode:
          type: string
          description: User's referral code for inviting others
          example: PULL2024ABC
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp
          example: "2024-01-15T10:30:00.000Z"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-01T00:00:00.000Z"
    AuthResponse:
      type: object
      required:
        - success
        - data
        - timestamp
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          required:
            - user
            - token
          properties:
            user:
              $ref: "#/components/schemas/User"
            token:
              type: string
              description: JWT authentication token
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time

    # ========================================
    # Trading Schemas
    # ========================================
...
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Missing or invalid authorization header
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: "2024-01-15T10:30:00.000Z"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid request parameters
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: "2024-01-15T10:30:00.000Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: "2024-01-15T10:30:00.000Z"
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register new user
      description: Create a new user account with email and password
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                error:
                  code: USER_EXISTS
                  message: An account with this email already exists
                timestamp: "2024-01-15T10:30:00.000Z"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: Authenticate user with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
                timestamp: "2024-01-15T10:30:00.000Z"
        "403":
          description: Account suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                error:
                  code: ACCOUNT_SUSPENDED
                  message: Your account has been suspended. Please contact support.
                timestamp: "2024-01-15T10:30:00.000Z"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh token
      description: Get a new JWT token using an existing valid token
      operationId: refreshToken
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      expiresIn:
                        type: string
                        example: 30d
                  requestId:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Account inactive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Logout
      description: Invalidate the current session (client should discard token)
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Logged out successfully
                  requestId:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
  /api/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Request password reset
      description: Send a password reset email to the user (always returns success to prevent email enumeration)
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Password reset email sent (if account exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: If an account exists with this email, you will receive a password reset link shortly.
                  requestId:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        "400":
          $ref: "#/components/responses/ValidationError"

  # ========================================
  # Trading Endpoints
  # ========================================
...